987654321/fstr_project/requirements.txt
Django>=3.2,<4.0  
djangorestframework>=3.12,<4.0  
psycopg2>=2.9,<3.0  
python-decouple>=3.5,<4.0  
pytest>=6.2,<7.0  
pytest-django>=4.0,<5.0  